+++ derby.log
@@ -1,11 +1,11 @@
 ----------------------------------------------------------------
-Fri May 09 12:46:03 CEST 2025:
-Booting Derby version The Apache Software Foundation - Apache Derby - 10.16.1.1 - (1901046): instance a816c00e-0196-b4a6-f68a-000001bd0460 
+Sun May 11 16:37:17 CEST 2025:
+Booting Derby version The Apache Software Foundation - Apache Derby - 10.16.1.1 - (1901046): instance a816c00e-0196-bfc7-636e-000004c590d0 
 on database directory C:\Users\Vandp\Desktop\Universitet\GreenhouseDB with class loader jdk.internal.loader.ClassLoaders$AppClassLoader@69663380 
 Loaded from file:/C:/Users/Vandp/.m2/repository/org/apache/derby/derby/10.16.1.1/derby-10.16.1.1.jar
 java.vendor=Eclipse Adoptium
 java.runtime.version=22.0.2+9
-user.dir=C:\Users\Vandp\Smart-Grid-Appliance
+user.dir=D:\Github\Smart-Grid-Appliance
 os.name=Windows 11
 os.arch=amd64
 os.version=10.0
