+++ frontend/src/App.jsx
@@ -1,8 +1,9 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import MainLayoutDom from "./components/Sidebar/MainLayoutDom.jsx";
 import LoginSignUp from "./components/LoginSignUp/LoginSignUp.jsx";
-import DataPage from "./pages/DataPage.jsx";
-import MainPage from "./pages/MainPage.jsx";
+import DataPage from "./pages/data-Page.jsx";
+import MainPage from "./pages/main-Page.jsx";
+import EditPage from "./pages/edit-Page.jsx";
 
 function App() {
     return (
@@ -11,7 +12,8 @@ function App() {
                 <Route path="/" element={<MainLayoutDom />}>
                     <Route index element={<MainPage />} />
                     <Route path="data" element={<DataPage />} />
-                    <Route path="/login" element={<LoginSignUp />} />
+                    <Route path="login" element={<LoginSignUp />} />
+                    <Route path="edit" element={<EditPage />} />
                 </Route>
             </Routes>
         </BrowserRouter>
