--- frontend/src/pages/HomePage.jsx
@@ -1,3 +1,0 @@
-export default function HomePage() {
-    return <h1>Welcome to the Home Page!</h1>;
-}
