+++ frontend/src/pages/home-Page.jsx
@@ -1,0 +1,3 @@
+export default function HomePage() {
+    return <h1>Welcome to the Home Page!</h1>;
+}
